requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
psycopg2-binary
pydantic>=2.0
pydantic-settings
python-dotenv
requests
pandas
numpy
scikit-learn
cfbd
httpx
